requirements.txt
asgiref==3.5.0
django==4.0.4
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
sqlparse==0.4.2
tzdata==2022.1

